feat: 마이페이지 예약 내역 및 리뷰 수정 로직 수정 및 자잘한 스타일 수정

+++ components/cart/CartItem.tsx
@@ -1,13 +1,14 @@
-import { useAtom } from 'jotai'
+import { useAtom, useAtomValue, useSetAtom } from 'jotai'
 import Image from 'next/image'
 import React, { useEffect, useState } from 'react'
 import { cartProductState } from '../../store/cartProductState'
 import AmountCounter from '../common/AmountCounter'
-import { Button } from '../common/Button'
-
 import { CartReturnType } from './cartTypes'
 import { useDeleteCart, useUpdateCart } from './hooks/useCart'
 import { useGetPrice } from './hooks/useGetPrice'
+import { isMobileDisplay } from '../../store/isMobileDisplay'
+import { modalState } from '../../store/modalState'
+import { useRouter } from 'next/router'
 
 interface PropsType {
   cartList: CartReturnType
@@ -20,13 +21,24 @@ const CartItem = ({ cartList, isAllItemSelected }: PropsType) => {
   const [additionalQuantity, setAdditionalQuantity] = useState(additionalCount)
   const [isSelected, setIsSelected] = useState(false)
   const [selectedProduct, setSelectedProduct] = useAtom(cartProductState)
+  const isMobile = useAtomValue(isMobileDisplay)
   const { mutate: updateCartMutate } = useUpdateCart()
   const { mutate: deleteCartItem } = useDeleteCart()
   const { getDiscountedPrice } = useGetPrice()
+  const setModal = useSetAtom(modalState)
+  const router = useRouter()
 
   const onClickDeleteCartItem = () => {
-    setSelectedProduct((prev) => prev.filter((selected) => selected.productId !== productId))
-    deleteCartItem(id)
+    setModal({
+      isOpen: true,
+      title: '주문 취소',
+      content: '주문을 취소하시겠습니까?',
+      confirmFunc: () => {
+        setSelectedProduct((prev) => prev.filter((selected) => selected.productId !== productId))
+        deleteCartItem(id)
+        router.reload()
+      },
+    })
   }
 
   const onClickCheckBox = (clickedCartItem: CartReturnType) => {
@@ -88,13 +100,8 @@ const CartItem = ({ cartList, isAllItemSelected }: PropsType) => {
   ])
 
   return (
-    <div className="w-full xs:w-[300px] shadow-basic rounded-md flex gap-4 my-4 p-4">
-      {/**
-       * 체크 버튼 / 상품 사진 / 상품 이름 및 서브타이틀 / 상품 수량 / 추가 상품 수량 / 구매 하기 / 삭제하기
-       * 전체 구매, 전체 삭제
-       */}
-
-      <div className="flex items-center ml-2">
+    <div className="w-full shadow-basic rounded-md my-4 p-2 md:p-4">
+      <div className="flex items-center gap-x-2 md:gap-x-4">
         <input
           type="checkbox"
           className="w-[16px] h-[16px]"
@@ -108,8 +115,6 @@ const CartItem = ({ cartList, isAllItemSelected }: PropsType) => {
             })
           }
         />
-      </div>
-      <div className="flex items-center">
         <Image
           src={product.descImg[0] ? product.descImg[0] : '/foodiful.jpeg'}
           alt="상품 사진"
@@ -117,24 +122,62 @@ const CartItem = ({ cartList, isAllItemSelected }: PropsType) => {
           height={60}
           className="w-[120px] h-[120px] rounded-sm"
         />
-      </div>
-      <div className="flex flex-col w-[18%] grow">
-        <span className="font-bold">{product.name}</span>
-        <span className="text-textDisabled">
-          {product.subTitle.length > 15 ? `${product.subTitle.slice(0, 15)}..` : product.subTitle}
-        </span>
-        {product.discount ? (
-          <>
-            <span className="text-textDisabled line-through">
-              {product.price.toLocaleString()}원
-            </span>
-            <span>{getDiscountedPrice(product.price, product.discount).toLocaleString()}원</span>
-          </>
-        ) : (
-          <span> {product.price.toLocaleString()}원</span>
-        )}
 
-        <div className="flex items-center gap-x-4 grow mt-[10px]">
+        <div className="flex flex-col w-[18%] grow break-keep">
+          <span className="font-bold">{product.name}</span>
+          <span className="text-textDisabled">
+            {isMobile && product.subTitle.length > 15
+              ? `${product.subTitle.slice(0, 15)}..`
+              : product.subTitle}
+          </span>
+          {product.discount ? (
+            <>
+              <span className="text-textDisabled line-through">
+                {product.price.toLocaleString()}원
+              </span>
+              <span>{getDiscountedPrice(product.price, product.discount).toLocaleString()}원</span>
+            </>
+          ) : (
+            <span> {product.price.toLocaleString()}원</span>
+          )}
+
+          {!isMobile && (
+            <div className="flex items-center gap-x-4 grow mt-[10px]">
+              <div className="flex flex-col justify-center items-center ">
+                <span className="text-sm font-semibold text-main">상품 수량</span>
+                <AmountCounter
+                  minAmount={1}
+                  amount={productQuantity}
+                  setAmount={setProductQuantity}
+                  limit={product.limitQuantity}
+                  size="md"
+                />
+              </div>
+              <div className="flex flex-col justify-center items-center">
+                <span className="text-sm font-semibold text-main">추가 상품 수량</span>
+                <AmountCounter
+                  minAmount={0}
+                  amount={additionalQuantity}
+                  setAmount={setAdditionalQuantity}
+                  limit={productQuantity}
+                  size="md"
+                />
+              </div>
+            </div>
+          )}
+        </div>
+
+        <div className="flex items-start gap-2 mr-2">
+          <button
+            className="bg-white text-textDisabled hover:text-disabled break-keep"
+            onClick={onClickDeleteCartItem}
+          >
+            삭제
+          </button>
+        </div>
+      </div>
+      {isMobile && (
+        <div className="flex justify-center items-center gap-x-4 grow mt-[10px]">
           <div className="flex flex-col justify-center items-center ">
             <span className="text-sm font-semibold text-main">상품 수량</span>
             <AmountCounter
@@ -156,15 +199,7 @@ const CartItem = ({ cartList, isAllItemSelected }: PropsType) => {
             />
           </div>
         </div>
-      </div>
-      <div className="flex items-start gap-2 mr-2">
-        <button
-          className="bg-white text-textDisabled hover:text-disabled"
-          onClick={onClickDeleteCartItem}
-        >
-          삭제
-        </button>
-      </div>
+      )}
     </div>
   )
 }
